style(MessageRecordObserver): organize hierarchy of MessageRecordObserver
