+++ src/MessageRecordObserverBase.ts
@@ -1,11 +1,11 @@
 import * as Rx from 'rxjs';
-import { TestMessageValue } from '../TestMessageValue';
+import { TestMessageValue } from './TestMessageValue';
 
 /**
  * An Observer records Observable values emitted in form of TestMessageValue
  *
  */
-export interface MessageRecordObserver<T = string> extends Rx.Observer<T> {
+export interface MessageRecordObserverBase<T = string> extends Rx.Observer<T> {
   /**
    * Metadata of subscribed Observable, recorded value of next(), error(), complete() calls.
    */
